rewrote test for connected graphs.
It doesn't quite work for now. The reason is, that the implication
in the formula was ignored and needs to be implemented into the
closure formula and phi_con. It is now ignored, that the statement
can still be true eventhough the first half of the condition (closure)
is false. (In that case it even has to be true either way)